% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_class.R
\name{print.sb_df}
\alias{print.sb_df}
\title{print.sb_df}
\usage{
\method{print}{sb_df}(x, ...)
}
\arguments{
\item{x}{A smartabase attributes object}

\item{...}{Reserved for future use}
}
\value{
A cleaned message
}
\description{
Hides the custom attributes that we attach the objects
}
