% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_smartabase.R
\name{pull_smartabase}
\alias{pull_smartabase}
\title{pull_smartabase}
\usage{
pull_smartabase(
  form,
  ...,
  url = NULL,
  type = "event",
  download_attachment = FALSE,
  start_date = NULL,
  end_date = NULL,
  last = NULL,
  start_time = "00:00 AM",
  end_time = "11:59 PM",
  username = NULL,
  password = NULL,
  filter_user_key = NULL,
  filter_user_value = NULL,
  filter_data_key = NULL,
  filter_data_value = NULL,
  filter_data_condition = "equal_to",
  include_missing_user = FALSE,
  guess_col_type = TRUE,
  get_uuid = FALSE,
  cloud_mode = FALSE,
  last_sync_time = NULL,
  shiny_progress_code = NULL,
  dev_mode = FALSE
)
}
\arguments{
\item{form}{string: name of Smartabase form}

\item{url}{string: Smartabase url e.g. 'example.smartabase.com/site'}

\item{type}{string: either 'event', 'profile' or 'synchronise'}

\item{download_attachment}{boolean: if TRUE, will download any files that
are included in the form via a file upload/multiple file upload field}

\item{start_date}{string: 'dd/mm/yyyy'}

\item{end_date}{string: 'dd/mm/yyyy'}

\item{username}{string: Smartabase username -- ignore if setup with
save_credentials()}

\item{password}{string: Smartabase password -- ignore if setup with
save_credentials()}
}
\value{
Smartabase data
}
\description{
Downloads data from a Smartabase event or profile form
}
\details{
This function pulls data from a Smartabase event form or profile form using
the Smartabase API. For more details see the help vignette:
\code{vignette("pulling-data", package = "smartabaseR")}
}
\examples{
\dontrun{
# Get one week of wellness data from example.smartabase.com/site:
wellness_data <- pull_smartabase(
  url = "example.smartabase.com/site",
  form = "Daily Wellness",
  username = "john.smith",
  password = "examplePassword",
  start_date = "15/04/2019",
  end_date = "22/04/2019"
)
}

}
\keyword{internal}
