% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_credentials}
\alias{save_credentials}
\title{save_credentials}
\usage{
save_credentials()
}
\value{
.Renviron
}
\description{
Helper that opens .Renviron file for storing credentials. Deprecated: only
used to support other deprecated exported functions like \code{\link[=pull_smartabase]{pull_smartabase()}}
and \code{\link[=push_smartabase]{push_smartabase()}}
}
\examples{
\dontrun{
save_credentials()
}
}
