% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{sb_delete_event}
\alias{sb_delete_event}
\title{Export event data}
\usage{
sb_delete_event(
  event_id,
  url,
  username,
  password,
  ...,
  option = sb_delete_event_option()
)
}
\arguments{
\item{event_id}{IDs of Smartabase events to be deleted}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options accessible via \code{\link[=sb_delete_event_option]{sb_delete_event_option()}} object}
}
\value{
Success/fail message
}
\description{
\code{sb_delete_event()} deletes data from a Smartabase event form using
the Smartabase API. It requires the user to input a valid Smartabase
event ID, url and credentials.

Please see \code{vignette("deleting-data", package = "smartabaseR")} for more
details.
}
\examples{
\dontrun{
# Delete one record with event_id = 999 from example.smartabase.com/site:
sb_delete_event(
  event_id = 999,
  url = "example.smartabase.com/site",
  username = "john.smith",
  password = "examplePassword"
)
}

}
