% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{sb_delete_event_option}
\alias{sb_delete_event_option}
\title{Set option parameters for \code{\link[=sb_delete_event]{sb_delete_event()}}}
\usage{
sb_delete_event_option(interactive_mode = TRUE)
}
\arguments{
\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}
}
\value{
A list of options with class = "sb_delete_option"
}
\description{
Set option parameters for \code{\link[=sb_delete_event]{sb_delete_event()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_delete_event_option()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_group_option(guess_col_type = FALSE, interactive_mode = FALSE)
}
}
