% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_functions.R
\name{sb_disable_user_cache}
\alias{sb_disable_user_cache}
\title{Disable caching for \code{sb_get_user()} and \code{sb_login()}}
\usage{
sb_disable_user_cache()
}
\description{
Disable caching for \code{sb_get_user()} and \code{sb_login()}. Use this function when
you always want the most up to date user and/or login data.
}
