% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_functions.R
\name{sb_enable_user_cache}
\alias{sb_enable_user_cache}
\title{Enable caching for \code{sb_get_user()} and \code{sb_login()}}
\usage{
sb_enable_user_cache(timeout = 2 * 60 * 60)
}
\arguments{
\item{timeout}{Measured in seconds, the time it takes until the cache
invalidates (within a given R session). By default, the timeout is 2 hours
(2 * 60 * 60 seconds).}
}
\description{
Enable caching for \code{sb_get_user()} and \code{sb_login()}. This function is useful
when you are making multiple calls to \code{sb_get_user()} and/or \code{sb_login()} with
the same user filters. By enabling caching, you can avoid redudantly calling
out to Smartabase for the same user data.
}
