% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_filter.R
\name{sb_get_event_filter}
\alias{sb_get_event_filter}
\title{Set filter parameters for \code{\link[=sb_get_event]{sb_get_event()}}}
\usage{
sb_get_event_filter(
  user_key = NULL,
  user_value = NULL,
  data_key = NULL,
  data_value = NULL,
  data_condition = NULL,
  events_per_user = NULL
)
}
\arguments{
\item{user_key}{The type of user variable to filter by. The possible values
are \code{c("about", "username", "email", "group", "current_group")}.}

\item{user_value}{The specific user value to filter for e.g. if
\code{user_key = "username"}, then perhaps \code{user_value = "john.smith"}.}

\item{data_key}{The name of the Smartabase field to filter by.}

\item{data_value}{The specific value to filter for in the selected data key}

\item{data_condition}{Specify the condition you want to apply when filtering
i.e. "equal_to", "not_equal_to", "contains", "less_than", "greater_than",
"less_than_or_equal_to", "greater_than_or_equal_to".}

\item{events_per_user}{The maximum number of events to return per athlete,
ordered by most recent. If not specified or set to NULL, all events will be
retrieved for each athlete}
}
\value{
A list of filters with class = "sb_export_filter"
}
\description{
Set filter parameters for \code{\link[=sb_get_event]{sb_get_event()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_event_filter()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_event_filter(user_key = "group", user_value = "Example Group")
}
}
\seealso{
\code{\link[=sb_get_profile_filter]{sb_get_profile_filter()}}
\code{\link[=sb_get_user_filter]{sb_get_user_filter()}}
\code{\link[=sb_sync_event_filter]{sb_sync_event_filter()}}

Other export filter helpers: 
\code{\link{sb_get_profile_filter}()},
\code{\link{sb_get_user_filter}()},
\code{\link{sb_sync_event_filter}()}
}
\concept{export filter helpers}
