% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_option.R
\name{sb_get_group_option}
\alias{sb_get_group_option}
\title{Set option parameters for \code{\link[=sb_get_group]{sb_get_group()}}}
\usage{
sb_get_group_option(guess_col_type = TRUE, interactive_mode = TRUE)
}
\arguments{
\item{guess_col_type}{If TRUE, \code{\link[=sb_get_event]{sb_get_event()}} uses \code{\link[readr:type_convert]{readr::type_convert()}}
to guess the column types of the exported data frame. If FALSE, all columns
will come through as character.}

\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}
}
\value{
A list of options with class = "sb_export_option"
}
\description{
Set option parameters for \code{\link[=sb_get_group]{sb_get_group()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_group_option()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_group_option(guess_col_type = FALSE, interactive_mode = FALSE)
}
}
\seealso{
\code{\link[=sb_get_event_option]{sb_get_event_option()}}
\code{\link[=sb_sync_event_option]{sb_sync_event_option()}}
\code{\link[=sb_get_user_option]{sb_get_user_option()}}
\code{\link[=sb_get_profile_option]{sb_get_profile_option()}}

Other export option helpers: 
\code{\link{sb_get_event_option}()},
\code{\link{sb_get_license_audit_option}()},
\code{\link{sb_get_profile_option}()},
\code{\link{sb_get_user_option}()},
\code{\link{sb_sync_event_option}()}
}
\concept{export option helpers}
