% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_license_audit.R
\name{sb_get_license_audit}
\alias{sb_get_license_audit}
\title{Export license audit data}
\usage{
sb_get_license_audit(login, url, username, password, ...)
}
\arguments{
\item{login}{Produced by \code{sb_login()}}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}
}
\value{
tibble: license audit data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Download license audit. Requires login produced \code{sb_login()}.
}
\examples{
\dontrun{
url <- "example.smartabase.com/example"
username <- "<admin_username>"
password <- "<admin_password>"

login <- sb_login(
  url = url,
  username = username,
  password = password
)

license_audit <- sb_get_license_audit(
  login = login,
  url = url,
  username = username,
  password = password
)

print(license_audit)
}
}
