% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_option.R
\name{sb_get_license_audit_option}
\alias{sb_get_license_audit_option}
\title{Set option parameters for \code{\link[=sb_get_license_audit]{sb_get_license_audit()}}}
\usage{
sb_get_license_audit_option(
  interactive_mode = TRUE,
  cache_audit = TRUE,
  cache_audit_timeout = 1800
)
}
\arguments{
\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}

\item{cache_audit}{Logical. If \code{TRUE}, the result of \code{\link[=sb_get_license_audit]{sb_get_license_audit()}}
will be cached on the local machine and reused within the current R session,
up to the timeout specified by \code{cache_audit_timeout}.
Caching is recommended because retrieving the license audit:
\itemize{
\item Can be computationally intensive on large sites
\item License audit data is slow-changing and unlike to change in a given
R session
}}

\item{cache_audit_timeout}{Numeric. Duration (in seconds) that a cached audit
result remains valid. After this timeout, the data will be refreshed from
the API. Note that cached results are only stored for the current R session.}
}
\value{
A list of options with class = "sb_export_option"
}
\description{
Set option parameters for \code{\link[=sb_get_license_audit]{sb_get_license_audit()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_license_audit_option()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_license_audit_option(cache_audit = TRUE)
}
}
\seealso{
\code{\link[=sb_get_event_option]{sb_get_event_option()}}
\code{\link[=sb_sync_event_option]{sb_sync_event_option()}}
\code{\link[=sb_get_user_option]{sb_get_user_option()}}
\code{\link[=sb_get_profile_option]{sb_get_profile_option()}}

Other export option helpers: 
\code{\link{sb_get_event_option}()},
\code{\link{sb_get_group_option}()},
\code{\link{sb_get_profile_option}()},
\code{\link{sb_get_user_option}()},
\code{\link{sb_sync_event_option}()}
}
\concept{export option helpers}
