% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sb_get_profile}
\alias{sb_get_profile}
\title{Export data from a Smartabase profile form}
\usage{
sb_get_profile(
  form,
  url,
  username,
  password,
  ...,
  filter = sb_get_profile_filter(),
  option = sb_get_profile_option()
)
}
\arguments{
\item{form}{Name of Smartabase profile form}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{filter}{More filters accessible via \code{\link[=sb_get_profile_filter]{sb_get_profile_filter()}}}

\item{option}{More options accessible via \code{\link[=sb_get_profile_option]{sb_get_profile_option()}}}
}
\value{
tibble: Smartabase profile data
}
\description{
This function exports data from a Smartabase profile form using the
Smartabase API. It requires the user to input a valid Smartabase
event form name, url and credentials.

For more details see the help vignette:
\code{vignette("exporting-data", package = "smartabaseR")}
}
\section{Filters}{

There are also a range extra filters than can be supplied to the \code{filter}
argument. In order to reduce argument clutter in \code{sb_get_profile()}, all
optional filters must be generated by the
\code{\link[=sb_get_profile_filter]{sb_get_profile_filter()}} function. Please see \code{\link[=sb_get_profile_filter]{sb_get_profile_filter()}}
for more details.
}

\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument. In order to reduce argument clutter in \code{sb_get_profile()}, all
options must be generated by the \code{\link[=sb_get_profile_option]{sb_get_profile_option()}} function. Please
see \code{\link[=sb_get_profile_option]{sb_get_profile_option()}} for more details.
}

\examples{
\dontrun{
# Get profile data from "Athlete Profile" form:
sb_get_profile(
  form = "Athlete Profile",
  url = "example.smartabase.com/site",
  username = "example.user",
  password = "examplePassword"
)
}

}
\seealso{
\code{\link[=sb_get_event]{sb_get_event()}}
\code{\link[=sb_get_user]{sb_get_user()}}
\code{\link[=sb_sync_event]{sb_sync_event()}}

Other export functions: 
\code{\link{sb_get_event}()},
\code{\link{sb_get_user}()},
\code{\link{sb_insert_event}()},
\code{\link{sb_sync_event}()}
}
\concept{export functions}
