% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_filter.R
\name{sb_get_profile_filter}
\alias{sb_get_profile_filter}
\title{Set filter parameters for \code{\link[=sb_get_profile]{sb_get_profile()}}}
\usage{
sb_get_profile_filter(
  user_key = c("user_id", "about", "username", "email", "group", "current_group"),
  user_value = NULL
)
}
\arguments{
\item{user_key}{The type of user variable to filter by. The possible values
are \code{c("about", "username", "email", "group", "current_group")}.}

\item{user_value}{The specific user value to filter for e.g. if
\code{user_key = "username"}, then perhaps \code{user_value = "john.smith"}.}
}
\value{
A list of filters with class = "sb_export_filter"
}
\description{
Set filter parameters for \code{\link[=sb_get_profile]{sb_get_profile()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_profile_filter()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_profile_filter(user_key = "group", user_value = "Example Group")
}
}
\seealso{
\code{\link[=sb_get_event_filter]{sb_get_event_filter()}}
\code{\link[=sb_sync_event_filter]{sb_sync_event_filter()}}
\code{\link[=sb_get_user_filter]{sb_get_user_filter()}}

Other export filter helpers: 
\code{\link{sb_get_event_filter}()},
\code{\link{sb_get_user_filter}()},
\code{\link{sb_sync_event_filter}()}
}
\concept{export filter helpers}
