% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_option.R
\name{sb_get_user_option}
\alias{sb_get_user_option}
\title{Set option parameters for \code{\link[=sb_get_event]{sb_get_event()}}}
\usage{
sb_get_user_option(
  interactive_mode = TRUE,
  include_all_cols = FALSE,
  include_missing_user = FALSE,
  guess_col_type = TRUE,
  cache_user = TRUE,
  cache_user_timeout = 1800
)
}
\arguments{
\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}

\item{include_all_cols}{If FALSE, only basic user information is included in
the returned data frame; e.g. user_id, username, about etc. If TRUE, extra
user information like phone, address and date of birth are also returned. If
only one user is returned, roles and groups information about that user is
included.}

\item{include_missing_user}{If TRUE, includes users who have not recorded
data in that form as rows of NA values.}

\item{guess_col_type}{If TRUE, \code{\link[=sb_get_event]{sb_get_event()}} uses \code{\link[readr:type_convert]{readr::type_convert()}}
to guess the column types of the exported data frame. If FALSE, all columns
will come through as character.}

\item{cache_user}{Logical. If \code{TRUE}, the result of \code{sb_get_user()}
will be cached on the local machine and reused within the current R session,
up to the timeout specified by \code{cache_user_timeout}. All of the export
functions (e.g. \code{sb_get_event()}, \code{sb_sync_event()} etc.) call
\code{sb_get_user()} internally, so caching is recommended given that retrieving
the user data can be
\itemize{
\item Can be computationally intensive on large sites
\item User data is slow-changing and unlike to change in a given R session
}}

\item{cache_user_timeout}{Numeric. Duration (in seconds) that cached user
data remains valid. After this timeout, the data will be refreshed from
the API. Note that cached results are only stored for the current R session.}
}
\value{
A list of options with class = "sb_export_option"
}
\description{
Set option parameters for \code{\link[=sb_get_event]{sb_get_event()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_user_option()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_user_option(guess_col_type = FALSE, interactive_mode = FALSE)
}
}
\seealso{
\code{\link[=sb_get_event_option]{sb_get_event_option()}}
\code{\link[=sb_sync_event_option]{sb_sync_event_option()}}
\code{\link[=sb_get_profile_option]{sb_get_profile_option()}}
\code{\link[=sb_get_group_option]{sb_get_group_option()}}

Other export option helpers: 
\code{\link{sb_get_event_option}()},
\code{\link{sb_get_group_option}()},
\code{\link{sb_get_license_audit_option}()},
\code{\link{sb_get_profile_option}()},
\code{\link{sb_sync_event_option}()}
}
\concept{export option helpers}
