% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{sb_login}
\alias{sb_login}
\title{Log into Smartabase using credentials}
\usage{
sb_login(url, username, password, ..., option = sb_get_user_option())
}
\arguments{
\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options accessible via \code{\link[=sb_login_option]{sb_login_option()}} object
calling \code{sb_login()}}
}
\value{
login object
}
\description{
\code{smartabaseR} uses this function internally to log the user in at the start
of each R session. We've exported this function since it can also be useful
when troubleshooting to ensure that your credentials are indeed valid.
}
\examples{
\dontrun{
sb_login(
  url = "example.smartabase.com/site",
  username = "john.smith",
  password = "example_password"
)
}
}
