% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sb_update_event}
\alias{sb_update_event}
\title{Update profile data}
\usage{
sb_update_event(
  df,
  form,
  url,
  username,
  password,
  ...,
  option = sb_update_event_option()
)
}
\arguments{
\item{df}{Data to be imported to Smartabase}

\item{form}{Name of Smartabase event form}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options accessible via \code{\link[=sb_update_event_option]{sb_update_event_option()}} object}
}
\value{
Data frame with extra details about import status
}
\description{
\code{sb_update_event()} imports data frame into a Smartabase event form using
the Smartabase API. It requires the user to input a data frame, a valid
Smartabase event form name, url and credentials.

If the data frame contains a profile record that already exists in
Smartabase, then that profile record is updated. The data frame must have an
\code{event_id} column that contains valid Smartabase event IDs.

For more details see the help vignette:
\code{vignette("importing-data")}
}
\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument, including options for setting a different ID column and supplying
the names of table fields, among others.  In order to reduce argument clutter
in \code{sb_update_event()}, all options must be generated by the
\code{\link[=sb_update_event_option]{sb_update_event_option()}} function. Please see \code{\link[=sb_update_event_option]{sb_update_event_option()}}
for more details.
}

