% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sb_upsert_event}
\alias{sb_upsert_event}
\title{Upsert event data}
\usage{
sb_upsert_event(
  df,
  form,
  url,
  username,
  password,
  ...,
  option = sb_insert_event_option()
)
}
\arguments{
\item{df}{Data to be imported to Smartabase}

\item{form}{Name of Smartabase event form}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options accessible via \code{\link[=sb_upsert_event_option]{sb_upsert_event_option()}} object}
}
\value{
Data frame with extra details about import status
}
\description{
\code{sb_upsert_event()} imports data frame into a Smartabase event form using
the Smartabase API. It requires the user to input a data frame, a valid
Smartabase event form name, url and credentials.

If the data frame contains an event that already exists in Smartabase, then
that event is updated. If the data frame contains an event does not exist
in Smartabase, then that event will be inserted as a new event.

For more details see the help vignette:
\code{vignette("importing-data", package = "smartabaseR")}
}
\section{start_date / end_date / start_time / end_time columns}{

When uploading data to Smartabase, the API requires metadata about when the
event started and stopped. \code{sb_upsert_event()} will first search for any
columns in the data named "start_time" and "end_time", which must have
values in h:mm AM or h:mm PM format.

If those columns do not exist, \code{start_time} and \code{end_time} will be set to
the current time and one hour after the current time, respectively.

Likewise, \code{sb_upsert_event()} will also search for any columns named
"start_date" or "end_date", which must have values formatted as dd/mm/YYYY.
If those columns do not exist, both \code{start_date} and \code{end_date} will be set
to the current date (unless the difference between \code{start_time} and
\code{end_time} spans midnight, in which case the \code{end_date} will be set to the
current date plus one day).
}

\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument, including options for setting a different ID column and supplying
the names of table fields, among others.  In order to reduce argument clutter
in \code{sb_upsert_event()}, all options must be generated by the
\code{\link[=sb_upsert_event_option]{sb_upsert_event_option()}} function. Please see \code{\link[=sb_upsert_event_option]{sb_upsert_event_option()}}
for more details.
}

