% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{sb_upsert_profile}
\alias{sb_upsert_profile}
\title{Upsert profile data}
\usage{
sb_upsert_profile(
  df,
  form,
  url,
  username,
  password,
  ...,
  option = sb_upsert_profile_option()
)
}
\arguments{
\item{df}{Data to be imported to Smartabase}

\item{form}{Name of Smartabase event form}

\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options accessible via \code{\link[=sb_upsert_profile_option]{sb_upsert_profile_option()}} object}
}
\value{
Data frame with extra details about import status
}
\description{
\code{sb_upsert_profile()} imports data frame into a Smartabase event form using
the Smartabase API. It requires the user to input a data frame, a valid
Smartabase event form name, url and credentials.

If the data frame contains a profile record that already exists in
Smartabase, then that profile record is updated. If the data frame contains
a profile record that does not exist in Smartabase, then that profile record
will be inserted as a new profile record.

For more details see the help vignette:
\code{vignette("importing-data")}
}
\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument, including options for setting a different ID column and supplying
the names of table fields, among others.  In order to reduce argument clutter
in \code{sb_upsert_profile()}, all options must be generated by the
\code{\link[=sb_upsert_profile_option]{sb_upsert_profile_option()}} function. Please see
\code{\link[=sb_upsert_profile_option]{sb_upsert_profile_option()}} for more details.
}

