% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_option.R
\name{sb_get_event_option}
\alias{sb_get_event_option}
\title{Set option parameters for \code{\link[=sb_get_event]{sb_get_event()}}}
\usage{
sb_get_event_option(
  interactive_mode = TRUE,
  include_user_data = TRUE,
  cache = TRUE,
  download_attachment = FALSE,
  include_missing_user = FALSE,
  guess_col_type = TRUE,
  include_uuid = FALSE
)
}
\arguments{
\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}

\item{include_user_data}{If TRUE, information like the athlete's full
name (i.e. about) is attached to the exported data frame. Setting this to
FALSE can improve query performance since \code{\link[=sb_get_event]{sb_get_event()}} doesn't need to
make an extra API call to grab the user data.}

\item{cache}{If TRUE, login token and user details are cached once per
session. Similar to \code{include_user_data}, caching can have performance
benefits when the user data set is large, such as on large Smartabase sites
with thousands of users. To invalidate the cache, restart your R session.}

\item{download_attachment}{If TRUE, will download any files that are
included in the form via a file upload field.}

\item{include_missing_user}{If TRUE, includes users who have not recorded
data in that form as rows of NA values.}

\item{guess_col_type}{If TRUE, \code{\link[=sb_get_event]{sb_get_event()}} uses \code{\link[readr:type_convert]{readr::type_convert()}}
to guess the column types of the exported data frame. If FALSE, all columns
will come through as character.}

\item{include_uuid}{If TRUE, each athlete's user-defined UUID
(universally unique identifier) is attached to the exported data frame.
Note, this is different to user_id, which is automatically generated by
Smartabase.}
}
\value{
A list of options with class = "sb_export_option"
}
\description{
Set option parameters for \code{\link[=sb_get_event]{sb_get_event()}}
}
\examples{
\dontrun{
# Calling the function with no arguments returns all the defaults
sb_get_event_option()

# Specifying specific arguments will alter those arguments alone while still
# returning the other defaults
sb_get_event_option(guess_col_type = FALSE, interactive_mode = FALSE)
}
}
\seealso{
\code{\link[=sb_sync_event_option]{sb_sync_event_option()}}
\code{\link[=sb_get_user_option]{sb_get_user_option()}}
\code{\link[=sb_get_profile_option]{sb_get_profile_option()}}
\code{\link[=sb_get_group_option]{sb_get_group_option()}}

Other export option helpers: 
\code{\link{sb_get_group_option}()},
\code{\link{sb_get_profile_option}()},
\code{\link{sb_get_user_option}()},
\code{\link{sb_sync_event_option}()}
}
\concept{export option helpers}
