% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sb_get_group}
\alias{sb_get_group}
\title{Export Smartabase group data}
\usage{
sb_get_group(url, username, password, ..., option = sb_get_group_option())
}
\arguments{
\item{url}{Smartabase url; e.g. 'example.smartabase.com/site'}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options generated by a \code{\link[=sb_get_user_option]{sb_get_user_option()}} object}
}
\value{
A tibble containing Smartabase group data
}
\description{
This function uses the Smartabase API to export the names of the
Smartabase groups that the logged in user has access to.

For more details see the help vignette:
\code{vignette("helper-functions", package = "smartabaseR")}
}
\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument. In order to reduce argument clutter in \code{sb_get_group()}, all
options must be generated by the \code{\link[=sb_get_group_option]{sb_get_group_option()}} function. Please
see \code{\link[=sb_get_group_option]{sb_get_group_option()}} for more details.
}

\examples{
\dontrun{
# Get user ID data:
group_data <- sb_get_group(
  url = "example.smartabase.com/site",
  username = "example.user",
  password = "examplePassword"
)
}
}
