% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sb_get_license_audit}
\alias{sb_get_license_audit}
\title{Export license audit data}
\usage{
sb_get_license_audit(
  url,
  username,
  password,
  ...,
  option = sb_get_license_audit_option()
)
}
\arguments{
\item{url}{Smartabase url e.g. "example.smartabase.com/site"}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{...}{These dots are for future extensions and must be empty}

\item{option}{More options generated by \code{\link[=sb_get_license_audit_option]{sb_get_license_audit_option()}}}
}
\value{
tibble: license audit data
}
\description{
Download license audit. Requires login produced by \code{sb_login()}. Logged-in
user must
}
\examples{
\dontrun{
url <- "example.smartabase.com/example"
username <- "<admin_username>"
password <- "<admin_password>"

license_audit <- sb_get_license_audit(
  url = url,
  username = username,
  password = password
)

print(license_audit)
}
}
