% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{sb_get_user}
\alias{sb_get_user}
\title{Export Smartabase user data}
\usage{
sb_get_user(...)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty}

\item{url}{Smartabase url; e.g. 'example.smartabase.com/site'}

\item{username}{Smartabase username}

\item{password}{Smartabase password}

\item{filter}{More filters generated by a \code{\link[=sb_get_user_filter]{sb_get_user_filter()}} object}

\item{option}{More options generated by a \code{\link[=sb_get_user_option]{sb_get_user_option()}} object}
}
\value{
A tibble containing Smartabase user data
}
\description{
\code{sb_get_user()} exports user data using the Smartabase API. It requires
the user to input a valid Smartabase url and valid Smartabase credentials.

For more details see the help vignette:
\code{vignette("helper-functions")}
}
\section{Filters}{

There are also a range extra filters than can be supplied to the \code{filter}
argument. In order to reduce argument clutter in \code{sb_get_user()}, all
optional filters must be generated by the
\code{\link[=sb_get_user_filter]{sb_get_user_filter()}} function. Please see \code{\link[=sb_get_user_filter]{sb_get_user_filter()}}
for more details.
}

\section{Options}{

There are also a range extra options than can be supplied to the \code{option}
argument. In order to reduce argument clutter in \code{sb_get_user()}, all
options must be generated by the \code{\link[=sb_get_user_option]{sb_get_user_option()}} function. Please
see \code{\link[=sb_get_user_option]{sb_get_user_option()}} for more details.
}

\examples{
\dontrun{
# Get user ID data about Jamie Anderson:
user_data <- sb_get_user(
  url = "example.smartabase.com/site",
  username = "example.username",
  password = "example_password",
  filter = sb_get_event_filter(
    user_key = "about",
    user_value = "Jamie Anderson"
  )
)

# Return all user columns:
user_data <- sb_get_user(
  url = "example.smartabase.com/site",
  username = "example.username",
  password = "example_password",
  option = sb_get_user_option(
    include_all_cols = TRUE
  )
)
}

}
\seealso{
\code{\link[=sb_get_profile]{sb_get_profile()}} for profile data.
\code{\link[=sb_get_user]{sb_get_user()}} for user data.
\code{\link[=sb_sync_event]{sb_sync_event()}} to synchronise with an event form.

Other export functions: 
\code{\link{sb_get_event}()},
\code{\link{sb_get_profile}()},
\code{\link{sb_sync_event}()}
}
\concept{export functions}
