% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_option.R
\name{sb_login_option}
\alias{sb_login_option}
\title{sb_login_option}
\usage{
sb_login_option(interactive_mode = TRUE)
}
\arguments{
\item{interactive_mode}{If TRUE, all messages are printed to the console.
If FALSE, they are suppressed. The idea is that \code{interactive_mode} should be
set to FALSE in automated environments to ensure logs aren't clogged up with
progress messages.}
}
\value{
list of named export options
}
\description{
Exports data from a Smartabase event form
}
