% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartabaseR-package.R
\docType{package}
\name{smartabaseR-package}
\alias{smartabaseR-package}
\alias{_PACKAGE}
\title{smartabaseR: R wrapper for the Smartabase API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R wrapper for the Smartabase API. Export, import or delete Smartabase data from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://teamworksapp.github.io/smartabaseR/}
  \item \url{https://github.com/Teamworksapp/smartabaseR}
}

}
\author{
\strong{Maintainer}: Zac Pross \email{zpross@teamworks.com}

Authors:
\itemize{
  \item James Day \email{jday@teamworks.com}
}

}
\keyword{internal}
